Moved the CartieFacade.php into a Facades folder and renamed it. Call LpMatrix\Cartie\Facades\Cartie to use
