+++ src/CartieServiceProvider.php
@@ -21,7 +21,7 @@ class CartieServiceProvider extends ServiceProvider
 
         if ($this->app->runningInConsole()) {
             $this->publishes([
-                __DIR__.'/../config/config.php' => config_path('cartie.php'),
+                __DIR__ . '/../config/config.php' => config_path('cartie.php'),
             ], 'config');
 
             // Publishing the views.
@@ -50,7 +50,7 @@ class CartieServiceProvider extends ServiceProvider
     public function register()
     {
         // Automatically apply the package configuration
-        $this->mergeConfigFrom(__DIR__.'/../config/config.php', 'cartie');
+        $this->mergeConfigFrom(__DIR__ . '/../config/config.php', 'cartie');
 
         // Register the main class to use with the facade
         $this->app->singleton('cartie', function () {
